chore: move steps to folder
